v0.8 Sprint 1 (QA-011): decompose CommunityHttpRequestProcessor — close GodClass + TooManyMethods
Extracts two responsibility-bundles from CommunityHttpRequestProcessor
into dedicated Community-internal helpers (package-private,
eu.exeris.kernel.community.http):

* CommunityHttpH2cUpgradeDetector — owns the *"is this an h2c upgrade
  or HTTP/2 prior-knowledge connection? if yes, transition to the
  HTTP/2 frame loop"* workflow. Instance class with constructor-injected
  HttpConfig + CommunityHttp2SessionProcessor; bundles the three
  decision methods (`isH2cEnabled`, `shouldHandlePriorKnowledge`,
  `handleHttp1UpgradeToH2c`) plus four pure helpers
  (`supportsHttp2`, `isHttp2PriorKnowledgePreface`, `isH2cUpgradeIntent`,
  `containsCsvTokenIgnoreCase`) plus the HTTP/2 connection-preface
  constant and the four header-token literals. 8 methods + 5 constants
  moved out.

* CommunityHttpAggregateTelemetry — owns the Phase 1C v0.6 aggregate
  buffer telemetry: *"how long has this aggregate been held, what
  fraction of requests were pipelined, should we force-release it?"*.
  Static utility, single entry point `applyAndRelease(state,
  maxAggregateBytes)`. 7 methods + 2 thresholds moved out.

Processor shrinks 408 → 258 LOC (−150). PMD class-level suppression
block:

    Before: {AvoidCatchingGenericException, CloseResource,
             CyclomaticComplexity, GodClass, TooManyMethods}
    After:  {AvoidCatchingGenericException, CloseResource,
             CyclomaticComplexity}

The remaining three suppressions are documented in the class header
with their root cause:

* `AvoidCatchingGenericException` — the request loop catches a generic
  `RuntimeException` to surface stream-close races as a clean return;
  `isExpectedStreamClose` carries the testable predicate.
* `CloseResource` — the aggregate `LoanedBuffer` is the flyweight that
  the keep-alive loop reuses across iterations; the try-with-resources
  on `TransportStream` + `ProcessingState` is the correct lifecycle.
* `CyclomaticComplexity` — `processIteration` + `readRequest`
  branching on keep-alive / aggregate-capacity / HTTP-2-upgrade is
  intrinsic to the codec contract; further decomposition would
  fragment the request lifecycle.

GodClass + TooManyMethods are gone — QA-011 exit criterion met.

Test surface preserved: `CommunityHttpRequestProcessorTest` calls
that previously asserted `isHttp2PriorKnowledgePreface` /
`isH2cUpgradeIntent` directly on the processor now call them on
`CommunityHttpH2cUpgradeDetector` (same package — both classes
package-private). Locked behaviour unchanged.

Local validation: full non-integration test sweep across all 10 modules
— BUILD SUCCESS. Sprint 1 cumulative PMD suppression delta after this
PR merges: −4 (out of the −5 sprint target across QA-010..014).
